Change weapons, and simple UI

+++ src/Components/HealthComponent.h
@@ -6,9 +6,9 @@
 #define HEALTHCOMPONENT_H
 
 struct HealthComponent {
-    int16_t health;
+    int32_t health;
 
-    explicit HealthComponent(const int16_t health = 100) : health(health) {}
+    explicit HealthComponent(const int32_t health = 100) : health(health) {}
 };
 
 #endif //HEALTHCOMPONENT_H
