+++ src/Components/ProjectileComponent.h
@@ -9,10 +9,10 @@
 
 struct ProjectileComponent {
     std::string tag;
-    int16_t damage;
-    int16_t selfDamage;
+    int32_t damage;
+    int32_t selfDamage;
 
-    explicit ProjectileComponent(const std::string& tag = "", const int16_t damage = 100, int16_t selfDamage = 100) {
+    explicit ProjectileComponent(const std::string& tag = "", const int32_t damage = 100, int32_t selfDamage = 100) {
         this->tag = tag;
         this->damage = damage;
         this->selfDamage = selfDamage;
